move source to src, adc .sequelizerc
