+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "api-jwt",
   "version": "1.0.0",
-  "main": "./api/server.js",
+  "main": "./src/api/server.js",
   "scripts": {
     "prestart": "npx sequelize db:drop && npx sequelize db:create && npx sequelize db:migrate && npx sequelize db:seed:all",
     "start": "nodemon .",
